backend/analytics-service/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
pydantic
python-multipart
requests
